+++ src/lint/index.js
@@ -2,8 +2,8 @@
 
 var linter = require('./linter');
 
-module.exports = function (path, cb) {
-	linter.start(path, function (error) {
+module.exports = function (path, ignorelist, cb) {
+	linter.start(path, ignorelist, function (error) {
 		if (error) {
 			return cb(error);
 		}