+++ src/lint/linter/src/linter.js
@@ -3,8 +3,8 @@
 var Lint = require('../lib/lint').Lint;
 var print = require('../lib/print');
 
-exports.start = function (fileList, cb) {
-	var lint = new Lint(fileList);
+exports.start = function (fileList, ignorelist, cb) {
+	var lint = new Lint(fileList, ignorelist);
 	lint.run(function (error) {
 		if (error) {
 			print.error(print.r(error.message));